--- farm_shane_fungi/farm_shane_fungi.routing.yml
@@ -1,7 +1,0 @@
-farm_shane_fungi.overview_page:
-  path: '/shanefungi/overviewpage'
-  defaults:
-    _controller: '\Drupal\farm_shane_fungi\Controller\ShaneFungiController::overviewPage'
-    _title: 'Testing Shane Fungi Overview Page.'
-  requirements:
-    _permission: 'access content'
