--- farm_shane_fungi/src/Controller/ShaneFungiController.php
@@ -1,23 +1,0 @@
-<?php
-namespace Drupal\farm_shane_fungi\Controller;
-
-use Drupal\Core\Controller\ControllerBase;
-
-/**
- * Provides route responses for the Example module.
- */
-class ShaneFungiController extends ControllerBase {
-
-  /**
-   * Returns a simple page.
-   *
-   * @return array
-   *   A simple renderable array.
-   */
-  public function overviewPage() {
-    return [
-      '#markup' => 'Hello, world',
-    ];
-  }
-
-}
