+++ src/Controller/ShaneFungiController.php
@@ -1,0 +1,23 @@
+<?php
+namespace Drupal\farm_shane_fungi\Controller;
+
+use Drupal\Core\Controller\ControllerBase;
+
+/**
+ * Provides route responses for the Example module.
+ */
+class ShaneFungiController extends ControllerBase {
+
+  /**
+   * Returns a simple page.
+   *
+   * @return array
+   *   A simple renderable array.
+   */
+  public function overviewPage() {
+    return [
+      '#markup' => 'Hello, world',
+    ];
+  }
+
+}
